Removing BucketTypeProps Operations
